First attempt at typescript. It's hacky, but better than nothing, and already found bugs.
