+++ package.json
@@ -4,6 +4,9 @@
   "private": true,
   "dependencies": {
     "@reach/router": "^1.2.1",
+    "@types/react": "^16.9.34",
+    "@types/react-dom": "^16.9.7",
+    "@types/uuid": "^7.0.3",
     "bootstrap": "^4.4.1",
     "bootstrap-social": "^5.1.1",
     "firebase": "^7.6.1",
@@ -48,7 +51,10 @@
     "fs-extra": "^8.1.0",
     "husky": "^4.2.1",
     "prettier": "^1.19.1",
-    "pretty-quick": "^2.0.1"
+    "pretty-quick": "^2.0.1",
+    "source-map-loader": "^0.2.4",
+    "ts-loader": "^7.0.1",
+    "typescript": "^3.8.3"
   },
   "husky": {
     "hooks": {
